// src/domains/AppotaPay/types/getPaymentTransaction.d.ts
// Generated by https://quicktype.io

import { AppotaPayError } from './appotaPayError';

export interface GetPaymentTransaction extends AppotaPayError {
    transaction: Transaction;
    partnerReference: PartnerReference;
}

export interface PartnerReference {
    order: Order;
}

export interface Order {
    id: string;
    info: string;
    extraData: string;
}

export interface Transaction {
    transactionId: string;
    reconciliationId: string;
    partnerCode: string;
    status: string;
    errorCode: number;
    errorMessage: string;
    orderAmount: number;
    amount: number;
    discountAmount: number;
    currency: string;
    bankCode: string;
    paymentMethod: string;
    action: string;
    fee: Fee;
    createdAt: Date;
    updatedAt: Date;
}

export interface Fee {
    customer_fee: number;
}