// src/domains/AppotaPay/types/newPayment.d.ts
// Generated by https://quicktype.io

import { AppotaPayError } from './appotaPayError';

export interface NewPaymentResponse extends AppotaPayError {
    transaction: Transaction;
    payment: Payment;
}

export interface Payment {
    url: string;
}

export interface Transaction {
    transactionId: string;
    status: string;
    errorCode: number;
    errorMessage: string;
    partnerCode: string;
    orderAmount: number;
    currency: string;
    bankCode: string;
    paymentMethod: string;
    action: string;
    createdAt: Date;
    updatedAt: Date;
}