// src/domains/AppotaPay/types/newCustomer.d.ts
// Generated by https://quicktype.io

import { AppotaPayError } from './appotaPayError';

export interface NewCustomerResponse extends AppotaPayError {
    customerRefId: string;
    customerId: string;
    firstName: string;
    lastName: string;
    mobileNumber: string;
    email: string;
    gender: null;
    dateOfBirth: string;
    placeOfBirth: string;
    nationality: string;
    createdAt: Date;
    updatedAt: Date;
}