// src/domains/AppotaPay/types/newPlan.d.ts
// Generated by https://quicktype.io

import { AppotaPayError } from './appotaPayError';

export interface NewPlanResponse extends AppotaPayError {
    planRefId: string;
    planId: string;
    customerId: string;
    currency: string;
    amount: number;
    paymentMethods: any[];
    immediateActionType: string;
    failedCycleAction: string;
    serviceName: string;
    status: string;
    actions: Action[];
    schedule: Schedule;
    notificationConfig: null;
    currencyExchange: null;
    createdAt: Date;
    updatedAt: Date;
}

export interface Action {
    url: string;
    action: string;
    method: string;
}

export interface Schedule {
    interval: string;
    intervalCount: number;
    totalRecurrence: number;
    anchorDate: Date;
    retryInterval: string;
    retryIntervalCount: number;
}