// src/domains/AppotaPay/types/getPlanCycles.d.ts
// Generated by https://quicktype.io

import { AppotaPayError } from './appotaPayError';

export interface GetPlanCyclesResponse extends AppotaPayError {
    data: Datum[];
    meta: Meta;
}

export interface Datum {
    cycleId: string;
    cycleNumber: number;
    currency: string;
    amount: number;
    scheduledAt: Date;
    status: string;
    createdAt: Date;
    updatedAt: Date;
}

export interface Meta {
    page: number;
    limit: number;
    total: number;
    pages: number;
}